+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## 0.9.0 (Unreleased)
 
+### Important Changes
+
+- Breaking: Experimental ACME support is added. Check the new configuration options and README.md for ACME support. Note that it is still under development and may have some issues.
+
+### Improvement
+
+- Refactor: lots of minor improvements
+- Deps
+
 ## 0.8.1
 
 ### Improvement